+++ frontend/src/container/TopNav/DateTimeSelection/config.ts
@@ -83,7 +83,7 @@ export const routesToSkip = [
 	ROUTES.ALERTS_NEW,
 	ROUTES.EDIT_ALERTS,
 	ROUTES.LIST_ALL_ALERT,
-	ROUTES.PIPELINES,
+	ROUTES.LOGS_PIPELINES,
 	ROUTES.BILLING,
 	ROUTES.SUPPORT,
 	ROUTES.WORKSPACE_LOCKED,
